requirements.txt
heroku3==4.2.3
Telethon==1.18.2
e-objects>=2.10.1
